+++ bot/config.js
@@ -16,6 +16,9 @@ module.exports = {
   "commands" : "http://franciscog.com/DerpyBot/commands/",
 
   "muted" : false,
+
+  // allow a user to give multiple props for the same song
+  "allow_multi_prop" : false,
   
   // turn on verbose logging
   'verboseLogging' : true, 
@@ -40,11 +43,20 @@ module.exports = {
 
   // make the bot save any song played (not skipped) to a playlist
   'saveSongs' : false, 
+  
+  // if a song is stuck because it has issues, automatically skip it
+  'autoskip_stuck' : true,
 
-  'autoskip' : {
-    'enabled' : true, // needs to be true for the following autoskips to be honored
-    'stuck' : true // auto skip stuck/unplayable songs
-  },
+  // should we warn the room that a song in the queue was played within X amount of hours
+  'recently_played_warning' : true,
+  // How many HOURS ago, or more, it's ok to allow a song to repeat
+  'recently_played_limit' : 8,
+  
+  // reset all user prop/flow/etc points back to 0 at the beginning of each month
+  'reset_points' : true,
+
+  // whether the bot should announce to the room every hour who the currently monthly leaders are
+  'hourly_leader' : true,
 
   'longSongs' : {
     // should we warn the room that a song exceeds the max length?